+++ api/run-test-container.sh
@@ -5,6 +5,20 @@
 # Any changes you make here will be lost next time 'npm run reconfigure' runs.
 # To make changes, see: scripts/reconfigure.mjs
 #
+
+#
+# Need some extra work to ensure that:
+#   - There is no API container running
+#   - The morionet network is available so we can attach to it
+#   - The reporter inside our test container has permissions to write coverage output
+#
+docker rm -f api
+docker network create morionet
+sudo rm -rf ./api/coverage/*
+mkdir ./api/coverage/tmp
+sudo chown 2112:2112 ./api/coverage/tmp
+
+
 docker run \
   -it --rm \
   --name=api \
@@ -12,6 +26,7 @@ docker run \
   --label morio.service=api \
   --log-driver=journald \
   --log-opt labels=morio.service \
+  --network morionet \
   --network-alias api \
   --init \
  \
@@ -20,7 +35,7 @@ docker run \
   -v /home/jdecock/git/morio:/morio  \
   -l "traefik.enable=true"  \
   -l "traefik.docker.network=morionet"  \
-  -l "traefik.http.routers.api.rule=(PathPrefix(`/-/api`, `/downloads`))"  \
+  -l "traefik.http.routers.api.rule=(PathPrefix(\`/-/api\`, \`/downloads\`, \`/coverage\`))"  \
   -l "traefik.http.routers.api.priority=100"  \
   -l "traefik.http.routers.api.service=api"  \
   -l "traefik.http.routers.api.entrypoints=https"  \