+++ config/services/proxy.mjs
@@ -18,7 +18,7 @@ export const resolveServiceConfiguration = (store) => {
       // Name to use for the running container
       container_name: 'proxy',
       // Aliases to use on the docker network (used to add unit test alias)
-      aliases: PROD ? [] : ['unit.test.morio.it'],
+      aliases: !PROD ? ['unit.test.morio.it'] : [],
       // Image to run (different in dev)
       image: 'traefik',
       // Image tag (version) to run